fix(agents): break mcp-panel↔dialog import cycle
depcruise's no-circular rule was failing on main: agent-mcp-panel.tsx
imported the dialog component while mcp-tools-resources-dialog.tsx
imported the McpServer type back from the panel.

Hoist McpServer into a leaf mcp-types.ts module both files import. The
panel re-exports the type so existing consumers are unaffected.
